refactor: deduplicate doc preprocessing
